survey-analysis-backend/pyproject.toml
[project]
name = "survey-analysis-engine"
version = "0.1.0"
description = "AI-Powered Survey Analysis Dashboard — Module 2 of 4"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.18",
    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    # Task Queue
    "celery[redis]==5.4.0",
    "redis==5.2.1",
    # LLM Gateway
    "litellm==1.55.3",
    # Data Analysis
    "pandas==2.2.3",
    "scipy==1.14.1",
    "numpy==2.1.3",
    # NLP
    "spacy==3.8.3",
    "textblob==0.18.0",
    # Serialization / Validation
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    # Environment Variables
    "python-dotenv==1.0.1",
    # Logging
    "structlog==24.4.0",
    "sentry-sdk[fastapi]==2.19.2",
    # WebSocket
    "python-socketio==5.11.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "import-linter==2.1",
    "radon==6.0.1",
    "ruff==0.8.4",
    "mypy==1.13.0",
    "httpx>=0.23.0,<0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "fitness: Architectural fitness function tests",
    "unit: Unit tests",
    "integration: Integration tests",
]

# ============================================================
# IMPORT LINTER — Fitness Function FF-01 & FF-03
# ============================================================
[tool.importlinter]
root_package = "src"

[[tool.importlinter.contracts]]
name = "Modules must not import each other's internals (FF-01)"
type = "forbidden"
source_modules = [
    "src.ingestion",
    "src.quality",
    "src.analytics",
    "src.visualization",
    "src.simulation",
    "src.chat_assistant",
]
forbidden_modules = [
    "src.ingestion.internals",
    "src.quality.internals",
    "src.analytics.internals",
    "src.visualization.internals",
    "src.simulation.internals",
    "src.chat_assistant.internals",
]
ignore_imports = [
    "src.ingestion.interfaces -> src.ingestion.internals",
    "src.quality.interfaces -> src.quality.internals",
    "src.analytics.interfaces -> src.analytics.internals",
    "src.visualization.interfaces -> src.visualization.internals",
    "src.simulation.interfaces -> src.simulation.internals",
    "src.chat_assistant.interfaces -> src.chat_assistant.internals",
]

[[tool.importlinter.contracts]]
name = "LLM Gateway Exclusivity (FF-03)"
type = "forbidden"
source_modules = [
    "src.ingestion",
    "src.quality",
    "src.analytics",
    "src.visualization",
    "src.simulation",
    "src.chat_assistant",
]
forbidden_modules = [
    "openai",
    "anthropic",
    "cohere",
]